expense-tracker-be/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
cffi==1.17.1
Django==4.2.16
django-allauth>=0.54.0
django-axes>=6.0.0
django-cors-headers>=4.0.0
django-environ>=0.11.0
django-filter>=23.0
django-ratelimit>=4.0.0
djangorestframework>=3.14.0
djangorestframework_simplejwt>=5.0.0
gunicorn==21.2.0
packaging==25.0
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.4
whitenoise==6.6.0
openpyxl==3.1.2
reportlab==4.2.5

